refactor: piece 테스트 수정
board.create()구조 변경에 따른 test코드 수정
