+++ src/test/java/chess/piece/KnightTest.java
@@ -38,7 +38,7 @@ class KnightTest {
                 new Knight(Position.of('a', '8'), Team.BLACK),
                 new Pawn(Position.of('b', '6'), Team.WHITE)
         );
-        Board board = Board.create(Pieces.from(pieces));
+        Board board = Board.create(Pieces.from(pieces), Turn.init());
         List<String> command = List.of("a8", "b6");
 
         assertDoesNotThrow(
@@ -50,7 +50,7 @@ class KnightTest {
     @DisplayName("나이트가 target위치로 진행할때 방해물이 있으면 넘어서 진행한다.")
     void moveKnightTest2() {
 
-        Board board = Board.create(Pieces.createInit());
+        Board board = Board.create(Pieces.createInit(), Turn.init());
         List<String> command = List.of("g1", "h3");
 
         assertDoesNotThrow(
@@ -65,7 +65,7 @@ class KnightTest {
                 new Knight(Position.of('a', '8'), Team.WHITE),
                 new Pawn(Position.of('c', '7'), Team.WHITE)
         );
-        Board board = Board.create(Pieces.from(pieces));
+        Board board = Board.create(Pieces.from(pieces), Turn.init());
         List<String> command = List.of("a8", "c7");
 
         assertThatThrownBy(
